Fix running with Python 3.10
(cherry picked from commit b647417e16f6146f4ae89608fd0494e7780da862)

+++ src/qtreactor.py
@@ -250,7 +250,7 @@ class QtReactor(posixbase.PosixReactorBase):
             timeout = 0
         else:
             timeout = self.timeout()
-        self._timer.setInterval(timeout * 1000)
+        self._timer.setInterval(int(timeout * 1000))
         self._timer.start()
 
     def runReturn(self, installSignalHandlers=True):
